Fix integer overflow by casting to unsigned as appropriate.
Recall the problem is that some variables (e.g., `uint16_t gfi->width`)
get promoted to int, and that very large GIFs might have signed 32-bit
sizes. The previous fix (just set gfi->width/height to unsigned) led
to lots of signedness warnings. We want to avoid the warnings. Really
what should be done is change all position variables to unsigned.
But I tried that, and it led to test failures. Scared of debugging the
failures, back off and fix the problem in a more targeted way.

Really a set of GIFs should be processed with clang's integer undefined
behavior sanitizer.

Conflicts:
	src/gifread.c
	src/quantize.c

+++ src/giffunc.c
@@ -398,7 +398,7 @@ Gif_CopyImage(Gif_Image *src)
   dest->interlace = src->interlace;
   if (src->img) {
     dest->img = Gif_NewArray(uint8_t *, dest->height + 1);
-    dest->image_data = Gif_NewArray(uint8_t, dest->width * dest->height);
+    dest->image_data = Gif_NewArray(uint8_t, (size_t) dest->width * (size_t) dest->height);
     dest->free_image_data = Gif_DeleteArrayFunc;
     if (!dest->img || !dest->image_data)
       goto failure;
@@ -789,7 +789,7 @@ Gif_SetUncompressedImage(Gif_Image *gfi, uint8_t *image_data,
 int
 Gif_CreateUncompressedImage(Gif_Image *gfi)
 {
-    size_t sz = gfi->width * gfi->height;
+    size_t sz = (size_t) gfi->width * (size_t) gfi->height;
     uint8_t *data = Gif_NewArray(uint8_t, sz ? sz : 1);
     return Gif_SetUncompressedImage(gfi, data, Gif_DeleteArrayFunc,
                                     gfi->interlace);
